+++ src/main/kotlin/sound/bins/frequency/dominant/frequency/DominantFrequencyCalculator.kt
@@ -1,4 +1,4 @@
-package dsp.bins.frequency.dominant.frequency
+package sound.bins.frequency.dominant.frequency
 
 import dsp.fft.FrequencyBins
