+++ src/test/kotlin/sound/bins/FrequencyBinToOctaveBinConverterIntegrationTests.kt
@@ -1,10 +1,10 @@
-package dsp.bins
+package sound.bins
 
 import dsp.fft.FrequencyBin
-import dsp.bins.notes.Notes
-import dsp.bins.octave.OctaveBin
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
+import sound.bins.notes.Notes
+import sound.bins.octave.OctaveBin
 
 class FrequencyBinToOctaveBinConverterIntegrationTests {
