+++ src/test/kotlin/sound/bins/frequency/dominant/frequency/TotalMagnitudeCalculatorTests.kt
@@ -1,24 +1,24 @@
-package dsp.bins.frequency.dominant.frequency
+package sound.bins.frequency.dominant.frequency
 
+import dsp.fft.FrequencyBin
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
-import dsp.fft.FrequencyBin
 
-class WeightedMagnitudeCalculatorTests {
+class TotalMagnitudeCalculatorTests {
 
     private val bin1 = FrequencyBin(10F, 20F)
     private val bin2 = FrequencyBin(20F, 30F)
     private val frequencyBins = listOf(bin1, bin2)
 
-    private fun createSUT(): WeightedMagnitudeCalculator {
-        return WeightedMagnitudeCalculator()
+    private fun createSUT(): TotalMagnitudeCalculator {
+        return TotalMagnitudeCalculator()
     }
 
     @Test
-    fun `calculate the weighted magnitude of frequency bins`() {
+    fun `calculate the total magnitude of frequency bins`() {
         val sut = createSUT()
         val actual = sut.calculate(frequencyBins)
-        assertEquals(800F, actual)
+        assertEquals(50F, actual)
     }
 
 }
