+++ src/test/kotlin/sound/bins/frequency/filters/CrossoverFilterTests.kt
@@ -1,8 +1,8 @@
-package dsp.bins.frequency.filters
+package sound.bins.frequency.filters
 
+import dsp.fft.FrequencyBin
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
-import dsp.fft.FrequencyBin
 
 class CrossoverFilterTests {
