+++ src/test/kotlin/toolkit/monkeyTest/NextCrossover.kt
@@ -1,6 +1,6 @@
 package toolkit.monkeyTest
 
-import dsp.bins.frequency.filters.Crossover
+import sound.bins.frequency.filters.Crossover
 import kotlin.random.Random
 
 fun nextCrossover(): Crossover {
