+++ src/test/kotlin/toolkit/monkeyTest/NextOctaveBin.kt
@@ -1,6 +1,6 @@
 package toolkit.monkeyTest
 
-import dsp.bins.octave.OctaveBin
+import sound.bins.octave.OctaveBin
 import kotlin.random.Random
 
 fun nextOctaveBin(
